2014-08-15 user voice screen for replying to messages 2014-08-19 updating response message on UserVoice success screen 2014-08-25 support 64-bit 2014-09-16 DE132 & DE 138 2014-09-17 styling 2014-09-17 don't suggest articles when replying; replies to WH-initiated message opens UV ticket 2014-09-22 handle network errors 2014-09-28 iPhone 6

+++ Classes/UVUser.m
@@ -121,7 +121,7 @@
 
 - (id)updateProperties:(NSDictionary *)updates delegate:(id<UVModelDelegate>)delegate {
   
-  NSString *pathWithId = [NSString stringWithFormat:@"/users/%d.json", _userId];
+  NSString *pathWithId = [NSString stringWithFormat:@"/users/%ld.json", (long)_userId];
   NSString *path = [UVUser apiPath:pathWithId];
   NSDictionary *payload = @{};
   if (updates) {
